refactor: move code to ReactantCore
