+++ .buildkite/pipeline.yml
@@ -10,13 +10,20 @@ steps:
     plugins:
       - JuliaCI/julia#v1:
           version: "{{matrix.version}}"
-      - JuliaCI/julia-test#v1:
-          test_args: "--gpu"
       - JuliaCI/julia-coverage#v1:
           codecov: true
           dirs:
             - src
             - ext
+            - lib/ReactantCore/src
+    commands: |
+      julia --project=. -e 'println("--- :julia: Instantiating project")
+          using Pkg
+          Pkg.develop([PackageSpec(path=pwd()), PackageSpec(path="lib/ReactantCore")])'
+
+      julia --project=. -e 'println("--- :julia: Run Tests")
+          using Pkg
+          Pkg.test(; coverage="user")'
     agents:
       queue: "juliagpu"
       cuda: "*"
@@ -34,7 +41,7 @@ steps:
         command: |
           julia --project=benchmark -e 'println("--- :julia: Instantiating project")
               using Pkg
-              Pkg.develop([PackageSpec(path=pwd())])'
+              Pkg.develop([PackageSpec(path=pwd()), PackageSpec(path="lib/ReactantCore")])'
 
           julia --project=benchmark -e 'println("--- :julia: Run Benchmarks")
               include("benchmark/runbenchmarks.jl")'
@@ -59,7 +66,7 @@ steps:
         command: |
           julia --project=benchmark -e 'println("--- :julia: Instantiating project")
               using Pkg
-              Pkg.develop([PackageSpec(path=pwd())])'
+              Pkg.develop([PackageSpec(path=pwd()), PackageSpec(path="lib/ReactantCore")])'
 
           julia --project=benchmark -e 'println("--- :julia: Run Benchmarks")
               include("benchmark/runbenchmarks.jl")'
