+++ .github/workflows/CI.yml
@@ -84,8 +84,19 @@ jobs:
           julia --color=yes --project=deps -e 'using Pkg; Pkg.instantiate()'
           SDKROOT=`xcrun --show-sdk-path` julia --color=yes --project=deps deps/build_local.jl
           cp LocalPreferences.toml test/
-      - uses: julia-actions/julia-runtest@v1
-        # if: steps.buildpkg.outcome == 'success'
+      - name: "Install Dependencies and Run Tests"
+        run: |
+          import Pkg
+          Pkg.Registry.update()
+          # Install packages present in subdirectories
+          dev_pks = Pkg.PackageSpec[]
+          for path in ("lib/ReactantCore",)
+              push!(dev_pks, Pkg.PackageSpec(; path))
+          end
+          Pkg.develop(dev_pks)
+          Pkg.instantiate()
+          Pkg.test(; coverage="user")
+        shell: julia --color=yes --code-coverage=user --depwarn=yes --project=. {0}
         id: run_tests
         env:
           JULIA_PKG_SERVER_REGISTRY_PREFERENCE: eager
@@ -108,7 +119,11 @@ jobs:
       - run: |
           julia --color=yes --project=docs -e '
             using Pkg
-            Pkg.develop([PackageSpec(path=pwd()), PackageSpec("Reactant_jll")])
+            Pkg.develop([
+              PackageSpec(path=pwd()),
+              PackageSpec("Reactant_jll"),
+              PackageSpec(path="lib/ReactantCore")
+            ])
             Pkg.instantiate()'
         env:
           JULIA_PKG_SERVER_REGISTRY_PREFERENCE: eager
