+++ Project.toml
@@ -8,11 +8,10 @@ Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
 CEnum = "fa961155-64e5-5f13-b03f-caf6b980ea82"
 Downloads = "f43a241f-c20a-4ad4-852c-f6b1247861c6"
 Enzyme = "7da242da-08ed-463a-9acd-ee780be4f1d9"
-ExpressionExplorer = "21656369-7473-754a-2065-74616d696c43"
 Libdl = "8f399da3-3557-5675-b5ff-fb832c97cbdb"
-MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
 OrderedCollections = "bac558e1-5e72-5ebc-8fee-abe8a469f55d"
 Preferences = "21216c6a-2e73-6563-6e65-726566657250"
+ReactantCore = "a3311ec8-5e00-46d5-b541-4f83e724a433"
 Reactant_jll = "0192cb87-2b54-54ad-80e0-3be72ad8a3c0"
 Scratch = "6c6a2e73-6563-6170-7368-637461726353"
 
@@ -32,11 +31,10 @@ ArrayInterface = "7.10"
 CEnum = "0.4, 0.5"
 Downloads = "1.6"
 Enzyme = "0.13"
-ExpressionExplorer = "1"
-MacroTools = "0.5.13"
 NNlib = "0.9"
 OrderedCollections = "1"
 Preferences = "1.4"
+ReactantCore = "0.1"
 Reactant_jll = "0.0.22"
 Scratch = "1.2"
 Statistics = "1.10"
