+++ lib/ReactantCore/Project.toml
@@ -1,0 +1,13 @@
+name = "ReactantCore"
+uuid = "a3311ec8-5e00-46d5-b541-4f83e724a433"
+authors = ["William Moses <wmoses@mit.edu>", "Valentin Churavy <vchuravy@mit.edu>", "Sergio Sánchez Ramírez <sergio.sanchez.ramirez@bsc.es>", "Paul Berg <paul@plutojl.org>"]
+version = "0.1.0"
+
+[deps]
+ExpressionExplorer = "21656369-7473-754a-2065-74616d696c43"
+MacroTools = "1914dd2f-81c6-5fcd-8719-6d5c9610ff09"
+
+[compat]
+ExpressionExplorer = "1"
+MacroTools = "0.5.13"
+julia = "1.10"
