+++ src/TracedRNumber.jl
@@ -12,6 +12,10 @@ mutable struct TracedRNumber{T} <: RNumber{T}
     end
 end
 
+ReactantCore.is_traced(::TracedRNumber) = true
+
+new_traced_value(::TracedRNumber{T}) where {T} = TracedRNumber{T}((), nothing)
+
 Base.eltype(::Type{TracedRNumber{T}}) where {T} = T
 
 Base.getindex(a::TracedRNumber{T}) where {T} = a
